Don't limit ActiveSupport to v4.x
We use ActiveSupport for gzipping but we don't want to tie this gem to
active_support '~> 4.0' (and therefore Rails 4.x) so let's just hope
that future versions of ActiveSupport don't change their gzip interface.
Seems unlikely, but I guess we can cross that bridge later and
potentially copy & paste the gzip code inte web_fetch if we really need
to.