Decouple serialisation from Ruby
Do not store Ruby objects in shared storage. We do not need to store
deferreds in storage; we just need to specify a callback that will store
the result in shared storage with the corresponding UID so that any
WebFetch worker can return the result.

Add Redis storage back end (WIP).